hotfix: check error status on interceptor

+++ frontend/src/app/interceptors/auth.interceptor.ts
@@ -8,8 +8,10 @@ intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> 
     return next.handle(req).pipe(
         tap(() => {}),
         catchError( error => {
-            localStorage.clear();
-            location.reload(); // When the guard checks that there is no token, It redirects to login
+            if(error.status === 401) { //Unauthorized
+                localStorage.clear();
+                location.reload(); // When the guard checks that there is no token, It redirects to login
+            }
             return throwError(error.message)
         })
     );
